Refactor: Remove misleading "legacy" naming and clean up combinators
**Renamed modules to reflect actual purpose:**
- `legacy_orchestration.rs` → `orchestration.rs`
  - This is the **current, stable high-level API** (LexPipeline)
  - Not legacy - it's the user-facing lexer/parser selection API
  - Updated all comments to clarify it's for high-level orchestration

- Clarified `builder.rs` (Pipeline) purpose:
  - Low-level TokenStream transformation pipeline
  - For chaining StreamMappers (internal infrastructure)
  - Different abstraction level from LexPipeline

**Removed actual legacy code:**
- Deleted `parsers/reference/combinators.rs`
  - Was just a re-export shim for builders.rs
  - Inlined the 2 imports into parser.rs and document.rs
  - Removed module declaration from reference.rs

**Updated documentation:**
- pipeline/mod.rs: Clarified distinction between high/low-level APIs
- builder.rs: Emphasized it's internal transformation infrastructure
- orchestration.rs: Removed "legacy" language, emphasized current stable API

**Architecture clarified:**
- **LexPipeline** (orchestration): High-level - select lexer/parser combinations
- **Pipeline** (builder): Low-level - chain TokenStream transformations
- Both are current and serve different purposes

✅ All 381 tests passing
✅ Zero clippy warnings
